refactor: piebiten: move internal audio source code to a separate package
