+++ piebiten/internal/audio/buffer_darwin.go
@@ -1,6 +1,6 @@
 // Copyright 2025 Jacek Olszak
 // This code is licensed under MIT license (see LICENSE for details)
 
-package internal
+package audio
 
 const audioBufferSizeInSeconds = 0.04 // 40ms
