+++ piebiten/internal/audio/buffer_windows.go
@@ -1,6 +1,6 @@
 // Copyright 2025 Jacek Olszak
 // This code is licensed under MIT license (see LICENSE for details)
 
-package internal
+package audio
 
-const audioBufferSizeInSeconds = 0.06 // 60ms
+const audioBufferSizeInSeconds = 0.02 // 20ms
